split the app component for render perf
