+++ Composer/packages/client/src/components/AppComponents/MainContainer.tsx
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+
+import { main } from './styles';
+import { SideBar } from './SideBar';
+import { RightPanel } from './RightPanel';
+import { Assistant } from './Assistant';
+
+export const MainContainer = () => {
+  return (
+    <div css={main}>
+      <SideBar />
+      <RightPanel />
+      <Assistant />
+    </div>
+  );
+};
