+++ Composer/packages/client/src/components/AppComponents/RightPanel.tsx
@@ -1,0 +1,34 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+import { useRecoilValue } from 'recoil';
+import { forwardRef } from 'react';
+
+import { ErrorBoundary } from './../ErrorBoundary/ErrorBoundary';
+import { RequireAuth } from './../RequireAuth/RequireAuth';
+import Routes from './../../router';
+import { applicationErrorState, dispatcherState, projectIdState } from './../../recoilModel';
+import { rightPanel, content } from './styles';
+
+const Content = forwardRef<HTMLDivElement>((props, ref) => <div css={content} {...props} ref={ref} />);
+
+export const RightPanel = () => {
+  const applicationError = useRecoilValue(applicationErrorState);
+  const { setApplicationLevelError, fetchProjectById } = useRecoilValue(dispatcherState);
+  const projectId = useRecoilValue(projectIdState);
+  return (
+    <div css={rightPanel}>
+      <ErrorBoundary
+        currentApplicationError={applicationError}
+        fetchProject={() => fetchProjectById(projectId)}
+        setApplicationLevelError={setApplicationLevelError}
+      >
+        <RequireAuth>
+          <Routes component={Content} />
+        </RequireAuth>
+      </ErrorBoundary>
+    </div>
+  );
+};
