+++ Composer/packages/client/src/router.tsx
@@ -11,7 +11,7 @@ import { resolveToBasePath } from './utils/fileUtil';
 import { data } from './styles';
 import { NotFound } from './components/NotFound/NotFound';
 import { BASEPATH } from './constants';
-import { botOpeningState, projectIdState, dispatcherState } from './recoilModel';
+import { botOpeningState, projectIdState, dispatcherState, designPageLocationState } from './recoilModel';
 import { LoadingSpinner } from './components/LoadingSpinner/LoadingSpinner';
 
 const DesignPage = React.lazy(() => import('./pages/design/DesignPage'));
@@ -25,7 +25,7 @@ const BotCreationFlowRouter = React.lazy(() => import('./components/CreationFlow
 
 const Routes = (props) => {
   const botOpening = useRecoilValue(botOpeningState);
-
+  const designPageLocation = useRecoilValue(designPageLocationState);
   return (
     <div css={data}>
       <Suspense fallback={<LoadingSpinner />}>
@@ -39,6 +39,11 @@ const Routes = (props) => {
             noThrow
             from="/bot/:projectId/language-understanding"
             to="/bot/:projectId/language-understanding/all"
+          />
+          <Redirect
+            noThrow
+            from="/bot/:projectId/dialogs"
+            to={`/bot/:projectId/dialogs/${designPageLocation.dialogId}`}
           />
           <Redirect noThrow from="/bot/:projectId/publish" to="/bot/:projectId/publish/all" />
           <Redirect noThrow from="/" to={resolveToBasePath(BASEPATH, 'home')} />
