+++ Composer/packages/client/src/utils/pageLinks.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import formatMessage from 'format-message';
 
-export const topLinks = (projectId: string, openedDialogId: string) => {
+export const topLinks = (projectId: string) => {
   const botLoaded = !!projectId;
   let links = [
     {
@@ -13,7 +13,7 @@ export const topLinks = (projectId: string, openedDialogId: string) => {
       disabled: false,
     },
     {
-      to: `/bot/${projectId}/dialogs/${openedDialogId}`,
+      to: `/bot/${projectId}/dialogs`,
       iconName: 'SplitObject',
       labelName: formatMessage('Design'),
       exact: false,
